fix to update url target after following http redirects
The url target is updated after following all http redirects to the final url.
This fixes borken states where the crawler is not arware of it's current location
because of http redirects.

This fix is not optimal as it causes duplicate requests to the same url.
Additionally the path of the redirect is not tracable.